+++ src/Rendering/HardwareTracing/HardwareTracingRenderer.cpp
