+++ src/Rendering/PathTracing/PathTracingPipeline.cpp
