+++ src/Rendering/PathTracing/PathTracingPipeline.hpp
