+++ src/Rendering/PathTracing/PathTracingRenderer.hpp
