+++ src/Rendering/SoftwareModern/SoftwareModernRenderer.cpp
