+++ src/Rendering/SoftwareTracing/SoftwareTracingPipeline.cpp
