+++ src/Rendering/SoftwareTracing/SoftwareTracingRenderer.cpp
