+++ src/Rendering/SoftwareTracing/SoftwareTracingRenderer.hpp
